+++ GLView/Core/GLView.m
@@ -172,9 +172,8 @@
     //set up layer
     CAEAGLLayer *eaglLayer = (CAEAGLLayer *)self.layer;
     eaglLayer.contentsScale = [UIScreen mainScreen].scale;
-    eaglLayer.drawableProperties = [NSDictionary dictionaryWithObjectsAndKeys:
-                                    [NSNumber numberWithBool:NO], kEAGLDrawablePropertyRetainedBacking,
-                                    kEAGLColorFormatRGBA8, kEAGLDrawablePropertyColorFormat, nil];
+    eaglLayer.drawableProperties = @{kEAGLDrawablePropertyRetainedBacking: @NO,
+                                    kEAGLDrawablePropertyColorFormat: kEAGLColorFormatRGBA8};
     
     //create context
     _context = [[EAGLContext alloc] initWithAPI:kEAGLRenderingAPIOpenGLES1